requirements-agent.txt
requests>=2.31
python-dotenv>=1.0
pydantic>=2.6
libcst>=1.4
fastapi>=0.110
uvicorn[standard]>=0.27
streamlit>=1.34
